Moved PDS job status state to common package #1620
- moved pds job status state to common packages
- also removed dependency to sechub-adapter-pds
  for pds tools because no longer necessary.
