+++ sechub-commons-pds/src/main/java/com/mercedesbenz/sechub/commons/pds/data/PDSJobStatusState.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-package com.mercedesbenz.sechub.pds.job;
+package com.mercedesbenz.sechub.commons.pds.data;
 
 public enum PDSJobStatusState {
