+++ sechub-pds-tools/build.gradle
@@ -10,7 +10,6 @@ plugins {
 }
 dependencies {
 
-    implementation project(':sechub-adapter-pds')
     implementation project(':sechub-commons-pds')
     implementation project(':sechub-commons-archive')
     
@@ -28,12 +27,3 @@ jar {
         attributes 'Main-Class': 'com.mercedesbenz.sechub.pds.tools.PDSToolsCLI'
     }
 }
-
-//task buildPDSTools(type: Jar, dependsOn: build) {
-//    group 'sechub'
-//    description 'Builds the PDS tools as standalone executable jar'
-//    
-//    manifest {
-//        attributes 'Main-Class': 'com.mercedesbenz.sechub.pds.tools.PDSToolsCLI'
-//    }
-//}
