+++ sechub-pds/src/main/java/com/mercedesbenz/sechub/pds/execution/PDSExecutionJobInQueueStatusEntry.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.mercedesbenz.sechub.commons.model.LocalDateTimeDeserializer;
 import com.mercedesbenz.sechub.commons.model.LocalDateTimeSerializer;
-import com.mercedesbenz.sechub.pds.job.PDSJobStatusState;
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatusState;
 
 public class PDSExecutionJobInQueueStatusEntry {
