+++ sechub-pds/src/main/java/com/mercedesbenz/sechub/pds/job/PDSJobRepositoryCustom.java
@@ -3,6 +3,8 @@ package com.mercedesbenz.sechub.pds.job;
 
 import java.util.Optional;
 
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatusState;
+
 public interface PDSJobRepositoryCustom {
 
     Optional<PDSJob> findNextJobToExecute();
