+++ sechub-pds/src/test/java/com/mercedesbenz/sechub/pds/job/PDSGetJobStatusServiceTest.java
@@ -13,6 +13,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatusState;
 import com.mercedesbenz.sechub.pds.PDSNotFoundException;
 import com.mercedesbenz.sechub.test.junit4.ExpectedExceptionFactory;
