+++ sechub-pds/src/test/java/com/mercedesbenz/sechub/pds/job/PDSStreamContentUpdateCheckerTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatusState;
+
 class PDSStreamContentUpdateCheckerTest {
 
     private static final String RUNNING = "RUNNING";
