+++ sechub-pds/src/test/java/com/mercedesbenz/sechub/pds/job/PDSUpdateJobTransactionServiceTest.java
@@ -13,6 +13,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.mockito.ArgumentCaptor;
 
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatusState;
 import com.mercedesbenz.sechub.pds.PDSNotAcceptableException;
 import com.mercedesbenz.sechub.pds.PDSNotFoundException;
 import com.mercedesbenz.sechub.test.junit4.ExpectedExceptionFactory;
